using crane for spawning a local registry

+++ scripts/smoke.sh
@@ -5,7 +5,7 @@ set -o pipefail
 
 readonly PROGDIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 readonly BUILDERDIR="$(cd "${PROGDIR}/.." && pwd)"
-readonly BIN_DIR="${BUILDERDIR}/.bin"
+readonly OPTIONS_JSON="${BUILDERDIR}/scripts/options.json"
 
 # shellcheck source=SCRIPTDIR/.util/tools.sh
 source "${PROGDIR}/.util/tools.sh"
@@ -15,7 +15,10 @@ source "${PROGDIR}/.util/print.sh"
 
 function main() {
   local name token
+  local registryPort registryPid localRegistry setupLocalRegistry pushBuilderToLocalRegistry
   token=""
+  registryPid=""
+  setupLocalRegistry=""
 
   while [[ "${#}" != 0 ]]; do
     case "${1}" in
@@ -53,32 +56,45 @@ function main() {
     name="testbuilder"
   fi
 
-  local registryPort registryPid localRegistry imageName
+  tools::install "${token}"
 
-  local push_builder_to_local_registry
-  if [ -f "$(dirname "${BASH_SOURCE[0]}")/options.json" ]; then
-    push_builder_to_local_registry="$(jq -r '.push_builder_to_local_registry //false' "$(dirname "${BASH_SOURCE[0]}")/options.json")"
-  else
-    push_builder_to_local_registry="false"
+  if [[ -f $OPTIONS_JSON ]]; then
+    setupLocalRegistry="$(jq -r '.setup_local_registry //false' $OPTIONS_JSON)"
   fi
 
-  tools::install "${token}"
+  if [[ "${setupLocalRegistry}" == "true" ]]; then
+    registryPort=$(get::random::port)
+    registryPid=$(local::registry::start $registryPort)
+    localRegistry="127.0.0.1:$registryPort"
+    export REGISTRY_URL="${localRegistry}"
+  fi
+
+  # tests::run
+
+  local pushBuilderToLocalRegistry
+
+  if [ -f $OPTIONS_JSON ]; then
+    pushBuilderToLocalRegistry="$(jq -r '.push_builder_to_local_registry //false' $OPTIONS_JSON)"
+  else
+    pushBuilderToLocalRegistry="false"
+  fi
 
   builder::create "${name}"
 
-  if [ "${push_builder_to_local_registry}" == "true" ]; then
-    registryPort=$(get::random::port)
-    registryPid=$(local::registry::start "$registryPort")
-    localRegistry="127.0.0.1:$registryPort"
-    docker tag "$name" "$localRegistry/$name"
-    docker push "$localRegistry/$name"
-    imageName="$localRegistry/$name"
+  if [ "${pushBuilderToLocalRegistry}" == "true" ]; then
+    docker tag "$name" "$REGISTRY_URL/$name"
+    docker push "$REGISTRY_URL/$name"
+    imageName="$REGISTRY_URL/$name"
   else
     imageName="$name"
   fi
-  tests::run $imageName
 
-  kill $registryPid
+  echo "tests::run $imageName"
+  # tests::run $imageName
+
+  if [[ "${setupLocalRegistry}" == "true" ]]; then
+    kill $registryPid
+  fi
 }
 
 function usage() {
@@ -99,10 +115,11 @@ function tools::install() {
   token="${1}"
 
   util::tools::crane::install \
-    --directory "${BIN_DIR}"
+    --directory "${BUILDERDIR}/.bin" \
+    --token "${token}"
 
   util::tools::pack::install \
-    --directory "${BIN_DIR}" \
+    --directory "${BUILDERDIR}/.bin" \
     --token "${token}"
 }
 
@@ -129,7 +146,7 @@ function image::pull::lifecycle() {
 
 function tests::run() {
   local name
-  name="$1"
+  name="${1}"
 
   util::print::title "Run Builder Smoke Tests"
